test: integration tests to separate folder
