+++ apps/backend/jest.config.mjs
@@ -1,7 +1,6 @@
 export default {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  maxWorkers: 1,
   roots: ['<rootDir>/src'],
   testMatch: ['**/__tests__/**/*.ts', '**/?(*.)+(spec|test).ts'],
   transform: {
